Extract General page to component
